Rearanging code
