+++ post/index.html
@@ -85,13 +85,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous">
         </script>
-
-    <script src="/src/js/handlers/getPosts.js" defer type="module"></script>
-    <script src="/src/js/handlers/createPost.js" defer type="module"></script>
-    <script src="/src/js/handlers/getComments.js" defer type="module"></script>
-
-
-
+    <script src="/src/js/index.js" type="module"></script>
 </body>
 
 </html>
