+++ posts/index.html
@@ -100,7 +100,6 @@
 
       <div class="container-fluid" id="allPosts">
 
-        <!--List of post thumbnails-->
 
       </div>
 
@@ -116,9 +115,7 @@
     integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous">
     </script>
 
-  <script src="/src/js/api/posts/delete.js" type="module"></script>
 
-  <script src="/src/js/handlers/search.js" type="module"></script>
 
   <script src="/src/js/index.js" type="module"></script>
