+++ src/js/api/posts/index.js
@@ -1,4 +1,4 @@
-export * from "./create.js";
-export * from "./read.js";
-export * from "./update.js";
-export * from "./delete.js";
+export * from "../posts/create.js";
+export * from "../posts/read.js";
+export * from "../posts/update.js";
+export * from "../posts/delete.js";
