+++ src/js/handlers/index.js
@@ -1,7 +1,8 @@
 //Barrel file 
-export * from "./login.js";
-export * from "./register.js";
-export * from "./createPost.js";
-export * from "./updatePost.js";
-export * from "./deletePost.js";
-export * from "./getPosts.js";
+export * from "./auth/login.js";
+export * from "./auth/register.js";
+export * from "./posts/createPost.js";
+export * from "./posts/updatePost.js";
+export * from "./posts/deletePost.js";
+export * from "./posts/getPosts.js";
+export * from "./posts/getComments.js"
