+++ src/js/handlers/posts/search.js
