--- src/js/templates/index.js
@@ -1,3 +1,0 @@
-//Barrel file 
-
-export * from "./postTemplate.js"
